# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WaitlistBypassCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "studio_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studio Name"
          }
        },
        "required": [
          "email",
          "name"
        ],
        "title": "WaitlistBypassCreateRequest",
        "type": "object"
      },
      "WaitlistEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "bypass_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bypass Token"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ip_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip Address"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rejection_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rejection Reason"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected",
              "invited"
            ],
            "title": "Status",
            "type": "string"
          },
          "studio_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studio Id"
          },
          "studio_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studio Name"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "studio_name",
          "status",
          "ip_address",
          "bypass_token",
          "reviewed_by",
          "reviewed_at",
          "rejection_reason",
          "studio_id",
          "created_at",
          "updated_at"
        ],
        "title": "WaitlistEntryResponse",
        "type": "object"
      },
      "WaitlistMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "WaitlistMessageRequest",
        "type": "object"
      },
      "WaitlistMessageResponse": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "sender_id": {
            "format": "uuid",
            "title": "Sender Id",
            "type": "string"
          },
          "waitlist_entry_id": {
            "format": "uuid",
            "title": "Waitlist Entry Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "waitlist_entry_id",
          "sender_id",
          "body",
          "created_at"
        ],
        "title": "WaitlistMessageResponse",
        "type": "object"
      },
      "WaitlistRejectRequest": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "WaitlistRejectRequest",
        "type": "object"
      },
      "WaitlistSubmitRequest": {
        "additionalProperties": false,
        "properties": {
          "captcha_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captcha Token"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "studio_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studio Name"
          }
        },
        "required": [
          "email",
          "name"
        ],
        "title": "WaitlistSubmitRequest",
        "type": "object"
      },
      "WaitlistSubmitResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "rejected",
              "invited"
            ],
            "title": "Status",
            "type": "string"
          },
          "studio_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studio Name"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "studio_name",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "WaitlistSubmitResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "DailyRiff API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/waitlist": {
      "get": {
        "description": "List all waitlist entries, optionally filtered by status.",
        "operationId": "list_waitlist_admin_waitlist_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "approved",
                    "rejected",
                    "invited"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WaitlistEntryResponse"
                  },
                  "title": "Response List Waitlist Admin Waitlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Waitlist",
        "tags": [
          "admin-waitlist"
        ]
      }
    },
    "/admin/waitlist/bypass": {
      "post": {
        "description": "Create a pre-approved waitlist entry with a bypass token for direct invites.",
        "operationId": "create_bypass_invite_admin_waitlist_bypass_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistBypassCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Bypass Invite",
        "tags": [
          "admin-waitlist"
        ]
      }
    },
    "/admin/waitlist/{entry_id}": {
      "get": {
        "description": "Get a single waitlist entry.",
        "operationId": "get_waitlist_entry_admin_waitlist__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "404": {
            "description": "Entry not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Waitlist Entry",
        "tags": [
          "admin-waitlist"
        ]
      }
    },
    "/admin/waitlist/{entry_id}/approve": {
      "post": {
        "description": "Approve a waitlist entry. Transitions status to 'approved'.",
        "operationId": "approve_waitlist_entry_admin_waitlist__entry_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "404": {
            "description": "Entry not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Waitlist Entry",
        "tags": [
          "admin-waitlist"
        ]
      }
    },
    "/admin/waitlist/{entry_id}/messages": {
      "get": {
        "description": "List messages for a waitlist entry.",
        "operationId": "list_waitlist_messages_admin_waitlist__entry_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WaitlistMessageResponse"
                  },
                  "title": "Response List Waitlist Messages Admin Waitlist  Entry Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "404": {
            "description": "Entry not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Waitlist Messages",
        "tags": [
          "admin-waitlist"
        ]
      },
      "post": {
        "description": "Send a message to a waitlist applicant.",
        "operationId": "send_waitlist_message_admin_waitlist__entry_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "404": {
            "description": "Entry not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Waitlist Message",
        "tags": [
          "admin-waitlist"
        ]
      }
    },
    "/admin/waitlist/{entry_id}/reject": {
      "post": {
        "description": "Reject a waitlist entry with optional reason.",
        "operationId": "reject_waitlist_entry_admin_waitlist__entry_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistRejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistEntryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad request \u2014 invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Superadmin role required"
          },
          "404": {
            "description": "Entry not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "422": {
            "description": "Validation error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Waitlist Entry",
        "tags": [
          "admin-waitlist"
        ]
      }
    },
...
    "/waitlist": {
      "post": {
        "description": "Submit a new waitlist entry from the marketing homepage.",
        "operationId": "submit_waitlist_waitlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WaitlistSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistSubmitResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Waitlist",
        "tags": [
          "waitlist"
        ]
      }
    }
  }
}